# user-service/internal/rest/api.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Users",
    "version": "1.0.0"
  },
...
  "basePath": "/",
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/api/users/{id}/change-email": {
      "put": {
        "summary": "Change a users Email",
        "operationId": "changeEmail",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/userspbChangeEmailResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/UsersServiceChangeEmailBody"
            }
          }
        ],
        "tags": [
          "User"
        ]
      }
    },
...
  },
  "definitions": {
    "UsersServiceChangeEmailBody": {
      "type": "object",
      "properties": {
        "email": {
          "type": "string"
        }
      }
    },
...
    "protobufAny": {
      "type": "object",
      "properties": {
        "@type": {
          "type": "string"
        }
      },
      "additionalProperties": {}
    },
    "rpcStatus": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "details": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/protobufAny"
          }
        }
      }
    },
...
    "userspbChangeEmailResponse": {
      "type": "object"
    },
...
  }
}